requirements.txt
# Core data science and analysis
numpy==1.26.4
pandas==2.2.1
matplotlib==3.8.3
seaborn==0.13.2

# Machine learning
scikit-learn==1.4.1.post1

# Web frameworks (uncomment if needed)
# flask==3.0.2
# fastapi==0.110.0
# requests==2.31.0

# Utilities
python-dotenv==1.0.1

#langgraph
langgraph
langchain-core
langchain-openai
langchain-anthropic
duckduckgo-search
langchain-community

# AI/LLM
anthropic

#iPython
ipython

#Tools
ddgs